requirements.txt
discord
python-dotenv
asyncio
typing